lint

+++ cpp/src/gandiva/precompiled/decimal_wrapper.cc
@@ -367,9 +367,11 @@ void castDECIMAL_utf8_internal(const char* in, int32_t in_length, int64_t out_pr
   int32_t scale_from_str;
   dec->FromString(std::string(in, in_length), dec.get(), &precision_from_str,
                   &scale_from_str);
-  gandiva::BasicDecimalScalar128 x({dec->high_bits(), dec->low_bits()}, precision_from_str, scale_from_str);
+  gandiva::BasicDecimalScalar128 x({dec->high_bits(), dec->low_bits()},
+                                   precision_from_str, scale_from_str);
   bool overflow = false;
-  auto out = gandiva::decimalops::Convert(x, out_precision, out_scale, &overflow);
+  auto out = gandiva::decimalops::Convert(x, static_cast<int32_t>(out_precision),
+                                          static_cast<int32_t>(out_scale), &overflow);
   *out_high = out.high_bits();
   *out_low = out.low_bits();
 }
